refactor(bench): extract BenchRecord into a shared bench utility
